refactor(asr): extract windowing and clean-window selection helpers
Move window start/weight/RMS helpers, clean_rawdata-style clean-window
selection + grid diagnostics, and the sample-mask/span utilities to
_windows.py. core.py re-exports them via __all__; no behavior change.
